+++ src/rasm/cpu.rs
@@ -1,105 +1,23 @@
 
 use super::{Instruction,AdrMode};
-use crate::DisplayStyle;
+
 use std::{
-    ops::{Index,IndexMut}
-    ,fmt::{Display,Formatter,Result},
     io::{Read,Write}
 };
 
-
-pub struct Memory {
-    min_addr : u16,
-    mem : Vec<i16>,
-    num_vars: usize,
-}
-
-impl Memory {
-    pub fn new(min_addr : u16,num_vars : usize) -> Self {
-        Memory {
-            min_addr,
-            mem : Vec::with_capacity(16),
-            num_vars,
-        }
-    }
-}
-
-impl Index<usize> for Memory {
-    type Output = i16;
-    fn index(&self, index: usize) -> &Self::Output {
-        if index as u16 > self.min_addr && self.min_addr >= 200 {
-            let i = index - self.min_addr as usize;
-            &self.mem[i + self.num_vars]
-        }else {
-            &self.mem[index]
-        }
-    }
-}
-impl IndexMut<usize> for Memory {
-    fn index_mut(&mut self, index: usize) -> &mut Self::Output {
-         if index as u16 > self.min_addr && self.min_addr >= 200 {
-            let i = index - self.min_addr as usize;
-            if i >= self.mem.len() {
-                self.mem.resize(i + self.num_vars,0);
-            }
-            
-            &mut self.mem[i + self.num_vars]
-        }else {
-            if index >= self.mem.len() {
-                self.mem.resize(index + 1, 0);
-            }
-            &mut self.mem[index]
-        }
-       
-    }
-}
+use super::mem::Memory;
 pub struct CPU<'a, I, O>
 where
     I : Read,
     O : Write {
     acc : i16,
     ix : i16,
-    pub pc : u16,
+    pc : u16,
     done : bool,
     flag_register : Flags,
     pub memory : &'a mut Memory, 
     display    : O,
     input      : I
-}
-pub struct Computer<'a,I : Read,O : Write> {
-    disp_style : DisplayStyle,
-    pub cpu        : CPU<'a,I,O>
-}
-pub struct ComputerBuilder<'a,I : Read,O : Write>  {
-    disp_style : Option<DisplayStyle>,
-    cpu        : Option<CPU<'a,I,O>>
-}
-impl<'a,I : Read,O : Write> ComputerBuilder<'a,I ,O > {
-    pub fn new() -> Self {
-
-        ComputerBuilder {
-            disp_style : None,
-            cpu  : None
-            
-        }
-    }
-    pub fn display_style(mut self,style : DisplayStyle) -> Self {
-        self.disp_style = Some(style);
-        self
-    }
-    pub fn attach_cpu(mut self, cpu : CPU<'a,I,O>) -> Self {
-        self.cpu = Some(cpu);
-        self
-    }
-    pub fn build(self) -> std::result::Result<Computer<'a,I,O>,&'static str> {
-        if self.cpu.is_none() || self.disp_style.is_none() {
-            return Err("Parts on CPU missing")
-        }
-       Ok( Computer {
-            cpu : self.cpu.unwrap(),
-            disp_style : self.disp_style.unwrap()
-        })
-    }
 }
 
 impl<'a,I,O> CPU<'a,I,O>
@@ -117,6 +35,18 @@ where
             display,
             input
         }
+    }
+    pub fn acc(&self) -> i16 {
+        self.acc
+    }
+    pub fn ix(&self) -> i16 {
+        self.ix
+    }
+    pub fn pc(&self) -> u16 {
+        self.pc
+    }
+    pub fn get_flag(&self,flag : FLAGS) -> bool {
+        self.flag_register.get_flag(flag)
     }
     pub fn done(&self) -> bool {
         self.done
@@ -267,95 +197,8 @@ where
 
 }
 
-impl<'a,I : Read,O : Write> Display for Computer<'a,I,O> {
-    fn fmt(&self, f: &mut Formatter<'_>) -> Result {
-    match self.disp_style{
-
-    DisplayStyle::Denary => {   write!(f, 
-r"
- ----------------------------------
-|              CPU                 |
-|    FLAGS                PC       |
-|    ---------           _______   |
-|   | N V Z C |         |{:>5}  |  |
-|   | {} {} {} {} |          -------   | 
-|    ---------                     |
-|    ACC                 IX        |
-|  ________          _________     |
-| |{:>8}|        |{:>8} |    |
-|  --------          ---------     |  
- ---------------------------------- ",
-                
-               self.cpu.pc,
-               self.cpu.flag_register.get_flag(FLAGS::N) as u8,
-               self.cpu.flag_register.get_flag(FLAGS::V) as u8,
-               self.cpu.flag_register.get_flag(FLAGS::Z) as u8,
-               self.cpu.flag_register.get_flag(FLAGS::C) as u8,
-               self.cpu.acc,
-               self.cpu.ix )
-
-
-    },
-    DisplayStyle::Binary => {
-       write!(f, 
-r"
- ---------------------------------------------------
-|                   CPU                             |
-|    FLAGS                          PC              |
-|    ---------                    _______           |
-|   | N V Z C |                  | {:^5} |          |
-|   | {} {} {} {} |                   -------           | 
-|    ---------                                      |
-|    ACC                               IX           |
-|  __________________          __________________   |
-| |{:#018b}|        |{:#018b}|  |
-|  ------------------          ------------------   |  
- --------------------------------------------------- ",
-                
-               self.cpu.pc,
-               self.cpu.flag_register.get_flag(FLAGS::N) as u8,
-               self.cpu.flag_register.get_flag(FLAGS::V) as u8,
-               self.cpu.flag_register.get_flag(FLAGS::Z) as u8,
-               self.cpu.flag_register.get_flag(FLAGS::C) as u8,
-               self.cpu.acc as u16,
-               self.cpu.ix as u16)
-
-
-
-    },
-    DisplayStyle::Hex => 
-    {
-       write!(f, 
-r"
- ----------------------------------
-|              CPU                 |
-|    FLAGS                PC       |
-|    ---------           _______   |
-|   | N V Z C |         |{:>5}  |  |
-|   | {} {} {} {} |          -------   | 
-|    ---------                     |
-|    ACC                 IX        |
-|  ________          _________     |
-| |{:#08x}|        |{:#08x} |    |
-|  --------          ---------     |  
- ---------------------------------- ",
-                
-               self.cpu.pc,
-               self.cpu.flag_register.get_flag(FLAGS::N) as u8,
-               self.cpu.flag_register.get_flag(FLAGS::V) as u8,
-               self.cpu.flag_register.get_flag(FLAGS::Z) as u8,
-               self.cpu.flag_register.get_flag(FLAGS::C) as u8,
-               self.cpu.acc as u16,
-               self.cpu.ix as u16)
-
-
-        }
-        
-      }
-   }
-}
 //Bit masks
-enum FLAGS {
+pub enum FLAGS {
     Z = 0b1000,
     N = 0b0100,
     V = 0b0010,
